feat: Change from aboslute to flex on hero
